+++ src/ansys/pre_commit_hooks/add_license_headers.py
@@ -32,6 +32,7 @@ import fileinput
 import json
 import os
 import pathlib
+from platform import python_version
 import shutil
 import sys
 from tempfile import NamedTemporaryFile
@@ -382,10 +383,17 @@ def add_hook_changes(before_hook: str, after_hook: str) -> None:
     before_hook_lines = before_hook_file.readlines()
     found_reuse_info = False
 
+    # Check if python version is 3.9 since fileinput.input()
+    # does not support the "encoding" keyword
+    if "3.9" in python_version():
+        file = fileinput.input(after_hook, inplace=True)
+    else:
+        file = fileinput.input(after_hook, inplace=True, encoding="utf8")
+
     # Copy file content before add-license-header was run into
     # the file after add-license-header was run.
     # stdout is redirected into the file if inplace is True
-    for line in fileinput.input(after_hook, inplace=True, encoding="utf8"):
+    for line in file:
         # Copy the new reuse lines into the file
         if _util.contains_reuse_info(line):
             count += 1