+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This document follows the conventions laid out in [Keep a CHANGELOG](https://keepachangelog.com/en/1.0.0).
 
+## [Unreleased]()
+
+### Fixed
+
+- Fix pytest python versions and fileinput [#118](https://github.com/ansys/pre-commit-hooks/pull/118)
+
 ## [0.2.6](https://github.com/ansys/pre-commit-hooks/releases/tag/v0.2.6) - January 11 2024
 
 ### Added